+++ src/lib.rs
@@ -728,7 +728,6 @@
 //!     Examples:
 //!
 //!     ```rust
-//!     # #![cfg_attr(feature = "iovec", feature(iovec))]
 //!     # #[cfg(feature = "transpose_methods")]
 //!     # use auto_enums::auto_enum;
 //!     # #[cfg(feature = "transpose_methods")]
@@ -816,8 +815,6 @@
 //! Note that support for these features are unstable and may cause incompatible changes between patch versions.
 //!
 //! * [`exact_size_is_empty`](https://github.com/rust-lang/rust/issues/35428) - Implements `ExactSizeIterator::is_empty`.
-//!
-//! * [`iovec`](https://github.com/rust-lang/rust/issues/58452) - Implements `io::Read::read_vectored` and `io::Write::write_vectored`.
 //!
 //! * [`read_initializer`](https://github.com/rust-lang/rust/issues/42788) - Implements `io::Read::read_initializer`.
 //!