+++ derive/src/enum_derive/mod.rs
@@ -160,6 +160,9 @@ lazy_static! {
             external::futures::async_read,
             #[cfg(feature = "futures")]
             external::futures::async_write,
+            // TODO
+            // #[cfg(feature = "futures")]
+            // external::futures::async_buf_read,
             // futures01
             #[cfg(feature = "futures01")]
             external::futures01::future,