+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Unreleased
 
+* Updated to new nightly. `iovec` stabilized. `#[enum_derive]` automatically detects the rustc version and supports `Read::read_vectored` and `Write::write_vectored` as the part of `Read` and `Write`.
+
 # 0.5.6 - 2019-04-16
 
-* Updated to new nightly.
+* Updated to new nightly. `futures_api` stabilized.
 
 # 0.5.5 - 2019-03-29